refactor: move the execution logic into `contract.rs`
The `state.rs` only provides the definition and interfaces for the
storage layer.
